pyproject.toml
[tool.poetry]
name = "camper"
version = "0.1.0"
description = ""
authors = ["hardiksuthar74 <hardiksuthar74@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.12"
uvicorn = "^0.34.0"
alembic = "^1.15.2"
sqlalchemy = "^2.0.40"
scalar-fastapi = "^1.0.3"
pydantic-settings = "^2.8.1"
asyncpg = "^0.30.0"
pydantic = {extras = ["email"], version = "^2.11.3"}
aiosmtplib = "^4.0.0"
cachetools = "^5.5.2"
redis = "^5.2.1"
asyncio = "^3.4.3"
aioredis = "^2.0.1"
pyjwt = "^2.10.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
